pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kinzie"
version = "1.0.0"
description = "Multi-strategy quant fund targeting Kalshi prediction market inefficiencies."
readme = "README.md"
requires-python = ">=3.11"
keywords = ["kalshi", "trading", "kelly-criterion", "arbitrage"]
dependencies = [
    "aiohttp>=3.9",
    "cryptography>=42.0",
    "python-dotenv>=1.0.0",
    "websockets>=14.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.22.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "ruff",
    "mypy",
    "hypothesis",
    "structlog",
]

[tool.hatch.build.targets.wheel]
packages = ["strategies", "core", "tools"]

[tool.ruff]
target-version = "py311"
line-length = 100
select = ["E", "F", "W", "I", "UP", "B"]
ignore = ["E501", "B008"]

[tool.ruff.per-file-ignores]
"tests/*" = ["S", "ANN"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
strict = false
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = ["core.*", "strategies.*"]
strict = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-q --tb=short"

[tool.coverage.run]
source = ["core", "strategies"]
omit = ["tests/*", "benchmarks/*", "research/*"]

[tool.coverage.report]
show_missing = true
skip_covered = false
